% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USCrashes-doc.R
\docType{data}
\name{USCrashes}
\alias{USCrashes}
\title{US Automobile Crashes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4019 rows and 8 columns.}
\usage{
USCrashes
}
\description{
US Automobile Crashes from FARS
(\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}).
}
\examples{
library(ggformula)
gf_line(deaths ~ date, data = USCrashes, color = ~ dayofweek)
gf_violinh(dayofweek ~ deaths, data = USCrashes)
}
\keyword{datasets}
